Sprint4/requirements.txt
-e .
aws-cdk.core==1.122.0
aws-cdk.aws_lambda==1.122.0
aws-cdk.aws_events==1.122.0
aws-cdk.aws_iam==1.122.0
aws-cdk.aws_events_targets==1.122.0
aws-cdk.aws_codepipeline==1.122.0
aws-cdk.aws_codepipeline_actions==1.122.0
aws-cdk.pipelines==1.122.0
pytest
boto3